requirements.txt
django-sheets==0.1.2
